rag/rag_open_source/rag_core/requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
docx2txt==0.8
fastapi==0.95.1
feedparser==6.0.11
Flask==3.0.2
Flask-Cors==4.0.0
html2text==2024.2.26
html_text==0.7.0
kafka-python==2.0.2
langchain==0.1.12
minio==7.2.7
numpy==1.23.5
openai==1.66.3
openpyxl==3.1.2
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.0
pymilvus==2.4.9
pymongo==4.10.1
PyMuPDF==1.20.2
pypdf==4.1.0
PyPDF2==3.0.1
pypinyin==0.48.0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2024.1
redis==5.2.0
requests==2.28.2
selenium==4.31.0
sse-starlette==2.0.0
tiktoken==0.8.0
tqdm==4.65.2
unstructured==0.11.2
urllib3==1.26.18
gunicorn==21.2.0
uvicorn==0.21.1
elasticsearch==8.15.0